# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Susan AI - Enterprise JARVIS-level Assistant API
  description: |
    A comprehensive enterprise-grade RESTful API system for Susan AI that supports all advanced features 
    and provides a robust foundation for a JARVIS-level assistant with enterprise capabilities.
    
    Core Features:
    - Streaming response support with Server-Sent Events and WebSocket
    - Advanced voice processing with batch operations and enhanced Whisper integration
    - Multimodal capabilities for images, documents, and file processing
    - Secure plugin architecture with sandboxing and runtime isolation
    - Vector-based memory system with semantic search and embedding management
    - Real-time collaboration with multi-user sessions and presence tracking
    - Comprehensive analytics and performance monitoring
    - Enterprise authentication with SSO, multi-tenancy, and audit logging
    - Advanced rate limiting and quota management
    - Scalable architecture supporting high-volume enterprise workloads
    
    Security Features:
    - Role-based access control (RBAC)
    - API key and JWT authentication
    - Request signing and validation
    - Audit logging and compliance tracking
    - Data encryption in transit and at rest
  version: 2.0.0
  contact:
    name: Susan AI Team
    email: support@susan-ai.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v2
...
    description: Production server
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - OAuth2: []
  - SignedRequest: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    OAuth2:
      type: oauth2
      description: OAuth2 authentication for enterprise SSO
      flows:
        authorizationCode:
          authorizationUrl: https://auth.susan-ai.com/oauth/authorize
          tokenUrl: https://auth.susan-ai.com/oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
            admin: Administrative access
            enterprise: Enterprise features access
    SignedRequest:
      type: apiKey
      in: header
      name: X-Signature
      description: Request signature for high-security operations
...
paths:
...
  /admin/system/logs:
    get:
      tags: [Administration]
      summary: Get system logs
      description: Retrieve system logs with filtering options
      parameters:
        - name: level
          in: query
          schema:
            type: string
            enum: [error, warn, info, debug]
        - name: service
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: System logs retrieved successfully
  /admin/config:
    get:
      tags: [Administration]
      summary: Get configuration
      description: Retrieve system configuration
      responses:
        '200':
          description: Configuration retrieved successfully

    put:
      tags: [Administration]
      summary: Update configuration
      description: Update system configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                configuration:
                  type: object
      responses:
        '200':
          description: Configuration updated successfully
...